fixed number of attacks modifier effect/added more bg2 effect opcodes
svn-id: r3855

+++ gemrb/plugins/Core/TileMap.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/TileMap.h,v 1.38 2006/02/22 18:38:20 avenger_teambg Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/TileMap.h,v 1.39 2006/04/12 20:32:10 avenger_teambg Exp $
  *
  */
 
@@ -59,7 +59,7 @@ public:
 	Door* GetDoor(Point &position);
 	Door* GetDoor(unsigned int idx);
 	Door* GetDoor(const char* Name);
-	unsigned int GetDoorCount() { return doors.size(); }
+	size_t GetDoorCount() { return doors.size(); }
 
 	/* type is an optional filter for container type*/
 	void AddContainer(Container *c);
@@ -68,20 +68,20 @@ public:
 	Container* GetContainer(unsigned int idx);
 	/* cleans up empty heaps, returns 1 if container removed*/
 	int CleanupContainer(Container *container);
-	unsigned int GetContainerCount() { return containers.size(); }
+	size_t GetContainerCount() { return containers.size(); }
 
 	InfoPoint* AddInfoPoint(const char* Name, unsigned short Type,
 		Gem_Polygon* outline);
 	InfoPoint* GetInfoPoint(Point &position);
 	InfoPoint* GetInfoPoint(const char* Name);
 	InfoPoint* GetInfoPoint(unsigned int idx);
-	unsigned int GetInfoPointCount() { return infoPoints.size(); }
+	size_t GetInfoPointCount() { return infoPoints.size(); }
 
 	TileObject* AddTile(const char* ID, const char* Name, unsigned int Flags, 
 		unsigned short* openindices, int opencount,unsigned short* closeindices, int closecount);
 	TileObject* GetTile(unsigned int idx);
 	TileObject* GetTile(const char* Name);
-	unsigned int GetTileCount() { return tiles.size(); }
+	size_t GetTileCount() { return tiles.size(); }
 
 	void AddOverlay(TileOverlay* overlay);
 	void DrawOverlays(Region screen);
